app/requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
pyyaml
torch
torchvision
opencv-python
prometheus-client